requirements.txt
fastapi
uvicorn
pymilvus
sentence-transformers
openai
requests
docker
python-dotenv